tweak(analytics): move google analytics bootstrap to _app.js
This will be globally initialized rather than a page by page basis.
