+++ pages/index.tsx
@@ -2,7 +2,6 @@ import dynamic from 'next/dynamic';
 import '@fontsource/montserrat/400.css';
 import '@fontsource/montserrat/500.css';
 import '@fontsource/montserrat/700.css';
-import Script from 'next/script';
 
 const Header = dynamic(() => import('../components/layout/Header'));
 const ZAPCTA = dynamic(() => import('../components/home/ZAPCTA'));
@@ -26,19 +25,6 @@ export default function Home() {
           </main>
         </div>
       </div>
-      <Script
-        src='https://www.googletagmanager.com/gtag/js?id=G-4DK3P3MG65'
-        strategy='afterInteractive'
-      />
-      <Script id='google-analytics' strategy='afterInteractive'>
-        {`
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){dataLayer.push(arguments);}
-          gtag('js', new Date());
-        
-          gtag('config', 'G-4DK3P3MG65');
-        `}
-      </Script>
     </>
   );
 }
